+++ src/components/SkillsToLearnSection/SkillsToLearnSection.js
@@ -1,0 +1,4 @@
+import React from 'react'
+import SkillsSection from '../SkillsSection/SkillsSection'
+
+export default (props) => <SkillsSection {...props} label='Add a skill you want to learn' placeholder='skills do you want to learn?' />