+++ src/routes/UserSettings/EditProfileTab/EditProfileTab.test.js
@@ -40,7 +40,7 @@ describe('EditProfileTab', () => {
   it('renders correctly', () => {
     const wrapper = shallow(<EditProfileTab currentUser={{}} />)
     expect(wrapper.find('Connect(UploadAttachmentButton)').length).toEqual(2)
-    expect(wrapper.find('SettingsControl').length).toEqual(7)
+    expect(wrapper.find('SettingsControl').length).toEqual(8)
     expect(wrapper.find('SocialControl').length).toEqual(3)
     expect(wrapper.find('Button').prop('color')).toEqual('gray')
     wrapper.setState({ changed: true })