+++ src/store/models/Person.js
@@ -1,6 +1,13 @@
 import PropTypes from 'prop-types'
 import { attr, fk, many, Model } from 'redux-orm'
 
+export const PersonSkillsToLearn = Model.createClass({})
+PersonSkillsToLearn.modelName = 'PersonSkillsToLearn'
+PersonSkillsToLearn.fields = {
+  person: fk('Person', 'personSkillsToLearn'),
+  skillToLearn: fk('Skill', 'personSkillsToLearn')
+}
+
 const Person = Model.createClass({
   toString () {
     return `Person: ${this.name}`
@@ -26,7 +33,13 @@ Person.fields = {
     to: 'Location',
     as: 'locationObject'
   }),
-  skills: many('Skill'),
+  skills: many({ to: 'Skill', as: 'skills', relatedName: 'peopleHaving' }),
+  skillsToLearn: many({
+    to: 'Skill',
+    relatedName: 'peopleLearning',
+    through: 'PersonSkillsToLearn',
+    throughFields: [ 'person', 'skillToLearn' ]
+  }),
   postsTotal: attr(),
   votesTotal: attr()
 }