+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to Hylo Evo (the Hylo front-end) will be documented in this 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Added Skills to Learn to user profiles
+
 ## [2.2.1] - 2020-12-05
 ### Added
 - Projects section to member profile